fix(oob): expose parseInvitation publicly (#834)
