+++ packages/core/src/index.ts
@@ -23,6 +23,7 @@ export type { Wallet } from './wallet/Wallet'
 export type { TransportSession } from './agent/TransportService'
 export { TransportService } from './agent/TransportService'
 
+import { parseInvitationUrl } from './utils/parseInvitation'
 import { uuid } from './utils/uuid'
 
 export * from './transport'
@@ -44,6 +45,7 @@ export * from './agent/Events'
 
 const utils = {
   uuid,
+  parseInvitationUrl,
 }
 
 export { utils }
